requirements.txt
transformers
sentence-transformers
faiss-cpu
uvicorn
fastapi
numpy
python-docx
PyPDF2
nltk
openai
